Deplyment 1

+++ apps/facturacion/models.py
@@ -10,9 +10,9 @@ class Venta(models.Model):
     nro = models.IntegerField(primary_key=True)
     dni_cliente = models.ForeignKey(Cliente,blank=True, null=True)
     fecha = models.DateField(default = timezone.now())
-    importe = models.IntegerField(blank=True, null=True)
-    saldo = models.IntegerField(blank=True, null=True)
-    total = models.IntegerField(blank=True, null=True)
+    importe = models.DecimalField(max_digits=10,decimal_places=2,blank=True, null=True)
+    saldo = models.DecimalField(max_digits=10,decimal_places=2,blank=True, null=True)
+    total = models.DecimalField(max_digits=10,decimal_places=2,blank=True, null=True)
     cancelado = models.BooleanField(default=True)
     observaciones = models.TextField(blank=True, null=True)
 
@@ -26,7 +26,7 @@ class Venta(models.Model):
 class DetalleVenta(models.Model):
     nro_venta = models.ForeignKey(Venta,blank=True)
     producto = models.ForeignKey(Producto,blank=True, null=True)
-    precio = models.IntegerField(default=0)
+    precio = models.DecimalField(max_digits=10,decimal_places=2,default=0.00)
     cantidad = models.IntegerField(default=0)
     def __unicode__(self):
         return "%s - %s" %(self.producto,self.cantidad)
@@ -35,13 +35,13 @@ class DetalleLente(models.Model):
     nro_venta = models.OneToOneField(Venta, blank=True)
     lente = models.ForeignKey(Lente)
     complementos = models.ManyToManyField(Aditivos)
-    precio = models.IntegerField(default=0)
+    precio = models.DecimalField(max_digits=10,decimal_places=2,default=0.00)
 
 class NotaPedido(models.Model):
     venta = models.ForeignKey(Venta)
     fecha = models.DateField(default=timezone.now())
-    importe = models.IntegerField(default=0)
-    saldo = models.IntegerField(default=0)
+    importe = models.DecimalField(max_digits=10,decimal_places=2,default=0.00)
+    saldo = models.DecimalField(max_digits=10,decimal_places=2,default=0.00)
 
     def __unicode__(self):
         return "Nota de Pedido nro. %i" %(self.id)
